+++ indispensable-josef/src/commonMain/kotlin/at/asitplus/signum/indispensable/josef/JwtPayload.kt
@@ -1,9 +1,8 @@
-package at.asitplus.signum.indispensable.josef.jwtpayload
+package at.asitplus.signum.indispensable.josef
 
 import at.asitplus.propigator.json.JsonBackingCodec
 import at.asitplus.propigator.json.JsonObjectBacked
 import at.asitplus.propigator.json.jsonSlice
-import at.asitplus.signum.indispensable.josef.JwtBaseClaims
 import at.asitplus.signum.indispensable.josef.io.joseCompliantSerializer
 import kotlinx.serialization.json.Json
 import kotlinx.serialization.json.JsonObject
