+++ indispensable-josef/src/commonMain/kotlin/at/asitplus/signum/indispensable/josef/jwtpayload/KeyAttestationPayload.kt
@@ -1,11 +1,10 @@
 package at.asitplus.signum.indispensable.josef.jwtpayload
 
 import at.asitplus.propigator.common.ObjectBackedValidated
-import at.asitplus.propigator.json.JsonBackingCodec
-import at.asitplus.propigator.json.JsonObjectBacked
 import at.asitplus.propigator.json.JsonObjectBackedSerializer
 import at.asitplus.propigator.json.jsonSlice
 import at.asitplus.signum.indispensable.josef.JwtBaseClaims
+import at.asitplus.signum.indispensable.josef.JwtPayload
 import at.asitplus.signum.indispensable.josef.io.joseCompliantSerializer
 import at.asitplus.signum.indispensable.josef.strictUnion
 import kotlinx.serialization.KSerializer
@@ -16,6 +15,7 @@ import kotlinx.serialization.json.JsonObject
 import kotlinx.serialization.json.encodeToJsonElement
 import kotlinx.serialization.json.jsonObject
 
+@Deprecated("Will move into VCK next release")
 @Serializable(with = KeyAttestationPayload.Serializer::class)
 data class KeyAttestationPayload(
     private val raw: JsonObject,
